bug 1114940 - rename worker-parent to worker
