+++ lib/sdk/content/worker.js
