+++ test/jetpack-package.ini
@@ -37,7 +37,7 @@ support-files =
 [test-content-script.js]
 [test-content-symbiont.js]
 [test-content-sync-worker.js]
-[test-content-worker-parent.js]
+[test-content-worker.js]
 [test-context-menu.js]
 [test-cortex.js]
 [test-cuddlefish.js]
